requirements.txt
openai==1.51.2
langchain==0.3.3
langchain-openai==0.2.2
langchain-pinecone==0.2.0
streamlit==1.39.0
docarray==0.40.0
pydantic==2.9.2
pytube==15.0.0
pytubefix==8.0.0
langchain-community==0.3.2
python-dotenv==1.0.1
tiktoken==0.8.0
pinecone-client==5.0.1
scikit-learn==1.5.2
ruff==0.6.9
google-api-core==2.21.0
ffmpeg==1.4
protoc-gen-openapiv2
scikit-learn
ruff
pypdf